fix: split appcompat.py's consumer-impact join out to fix the file-size hard cap
The merge combining ADR-067 C-S2's own appcompat.py growth (+24 lines) with
a concurrent PR's consumer-specification slice (+59 lines) pushed the file
to 2011 lines, tripping the AI-readiness gate's 2000-line hard cap (no
allowlist possible, unlike the reviewed no_growth baselines).

Per this file's own root-cause-not-instance convention, split responsibility
out rather than trim to fit: the whole consumer/source-graph join subsystem
(ADR-057 G29 Phase 4 -- _library_source_graph, consumer_impact_explanations,
_format_consumer_impact, _has_impact_evidence, enrich_covered_changes,
_merge_consumer_impact_paths, _choose_primary_match, attach_consumer_impact)
moves to a new sibling module, abicheck/appcompat_consumer_impact.py. It was
already the one self-contained subsystem in appcompat.py with exactly three
external call sites (all from scope_diff_to_app) and no dependency on the
rest of that file's ELF/PE/Mach-O parsing, disposition ledger, or CLI-facing
result types -- a real seam, not an arbitrary line cut.

appcompat.py: 2011 -> 1574 lines. Registered the new flat-root module in
architecture/modules.yaml (workflows legacy_paths + legacy_root_modules,
mirroring appcompat.py's own entries) and updated the handful of unit tests
that imported the moved private helpers directly.
